Battle v1

See merge request COMP2511/22T2/teams/F12A_ICECREAM/project!45

+++ src/main/java/dungeonmania/dynamic_entity/Spider.java
@@ -31,8 +31,8 @@ public class Spider extends DynamicEntity {
 
     public Spider(String id, Position xy, JSONObject config) {
         super(id, "spider", xy);
-        this.attack = config.getInt("spider_attack");
-        this.health = config.getInt("spider_health");
+        this.attack = config.getDouble("spider_attack");
+        this.health = config.getDouble("spider_health");
         this.direction = "clockwise";
         cyclePositions = generatePositions(this.getPosition());
     }
